+++ packages/core/src/mocks/Custom.svelte
@@ -1,7 +1,7 @@
 <script>
 	import { defaultFields } from '../../src/constants'
-	export let content
+	export let value
 	export let fields = defaultFields
 </script>
 
-<span>{content[fields.text]}</span>
+<span>{value[fields.text]}</span>