+++ packages/layout/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @rokkit/layout
 
+## 1.0.0-next.23
+
+### Patch Changes
+
+- Updated version with revamped Icon component.
+- Updated dependencies
+  - @rokkit/core@1.0.0-next.23
+  - @rokkit/markdown@1.0.0-next.23
+
 ## 1.0.0-next.22
 
 ### Patch Changes